requirements.txt
google-search-results-2.4.2
langchain-0.0.308
python-dotenv-1.0.0
openai-0.28.1